sprite event 處理從 Layer 移至 LayerSprite
順便調整 asWidget() 位置
